requirements.txt
django==2.1.9
djangorestframework
django-cors-headers==3.2.1
